entrenamiento-ia-seguros/requirements.txt
pandas
mysql-connector-python
scikit-learn
joblib
numpy
